Implement ops on `&Wrapping<T>`
