+++ src/libcore/lib.rs
@@ -100,6 +100,9 @@ use prelude::v1::*;
 #[macro_use]
 mod macros;
 
+#[macro_use]
+mod internal_macros;
+
 #[path = "num/float_macros.rs"]
 #[macro_use]
 mod float_macros;
